pyproject.toml
[tool.poetry]
name = "bash-scripts"
version = "0.1.0"
description = ""
authors = ["ezra-wick <ezra.wick@ya.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.1"
uvicorn = "^0.30.3"
Django = "^5.0.7"
channels = "^4.1.0"
channels-redis = "^4.2.0"
djangorestframework = "^3.15.2"
colorlog = "^6.8.2"
python-dotenv = "^1.0.1"
django-filter = "^24.2"
opencv-python = "^4.10.0"
numpy = "^2.0.1"
dlib = "^19.24.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
